requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-jinja2==1.6
aiosignal==1.3.2
attrs==25.1.0
click==8.1.8
ConsoleWidget @ git+https://github.com/AlviseDeFaveri/ConsoleWidget.git@c667aa9606ab02423a5a61b25d6f38632b41ab49
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
platformdirs==4.3.6
propcache==0.2.1
Pygments==2.19.1
pyverilator-mm @ git+https://github.com/bat52/pyverilator.git@0799cfa0d19222d8e746b72950aaa412530f3d06
rich==13.9.4
textual==2.0.3
textual-dev==1.7.0
textual-serve==1.1.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
yarl==1.18.3
